xtimer: Split header into API and implementation
